+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Integrate latest change from `Vaultwarden`/`SSO PR` up to [a71da2d](https://github.com/dani-garcia/vaultwarden/commit/a71da2d0a4e93034397134b17078a27829264043)
 - Use [oidc_web_vault](https://github.com/Timshel/oidc_web_vault) version `v2025.7.2-1`
+- Change to `SSO_ORGANIZATIONS_ALL_COLLECTIONS` to stop using a hidden field.
+  On user invitation will now add read only access to all available collection (Does not grant access to collections created afterwards.).
 
 # v2025.6.2-1